requirements.txt
transformers>=4.35.0
fasttext-wheel>=0.9.2
numpy>=1.24,<2.0
sentencepiece>=0.1.99
sacremoses>=0.0.53
tqdm>=4.65.0
requests>=2.31.0
llama-cpp-python